refactor(plugins): share package traversal and isolate test fixtures
