+++ nix/scripts/runtime-plugin-locks/package.mjs
@@ -1,56 +1,7 @@
 import fs from "node:fs";
 import path from "node:path";
+import { collectPackageRoots, packageRootForName } from "../runtime-plugin/package-tree.mjs";
 import { isRecord, optionalString, run } from "./io.mjs";
-
-function collectPackageRoots(nodeModulesDir, baseRel = "node_modules") {
-  if (!fs.existsSync(nodeModulesDir)) {
-    return [];
-  }
-
-  const roots = [];
-  for (const entry of fs.readdirSync(nodeModulesDir).sort()) {
-    if (entry === ".bin") {
-      continue;
-    }
-
-    const entryPath = path.join(nodeModulesDir, entry);
-    const entryRel = `${baseRel}/${entry}`;
-    if (!fs.lstatSync(entryPath).isDirectory()) {
-      continue;
-    }
-
-    if (entry.startsWith("@")) {
-      for (const scopedName of fs.readdirSync(entryPath).sort()) {
-        const scopedPath = path.join(entryPath, scopedName);
-        const scopedRel = `${entryRel}/${scopedName}`;
-        if (fs.lstatSync(scopedPath).isDirectory()) {
-          roots.push(scopedRel);
-          roots.push(...collectPackageRoots(path.join(scopedPath, "node_modules"), `${scopedRel}/node_modules`));
-        }
-      }
-    } else {
-      roots.push(entryRel);
-      roots.push(...collectPackageRoots(path.join(entryPath, "node_modules"), `${entryRel}/node_modules`));
-    }
-  }
-
-  return roots;
-}
-
-function packageRootForName(packageName) {
-  if (packageName.startsWith("@")) {
-    const parts = packageName.split("/");
-    if (parts.length !== 2 || !parts[0] || !parts[1]) {
-      throw new Error(`invalid scoped package name ${packageName}`);
-    }
-    const [scope, name] = parts;
-    return `node_modules/${scope}/${name}`;
-  }
-  if (!packageName || packageName.includes("/")) {
-    throw new Error(`invalid package name ${packageName}`);
-  }
-  return `node_modules/${packageName}`;
-}
 
 function declaredDependencyRoots(dependencies = {}, optionalDependencies = {}) {
   return [
